requirements.txt
anyio==3.6.1
APScheduler==3.6.3
beautifulsoup4==4.11.1
cachetools==4.2.2
certifi==2022.6.15
click==8.1.3
fastapi==0.79.0
gunicorn==20.1.0
h11==0.13.0
httptools==0.4.0
idna==3.3
pydantic==1.9.1
pymongo==4.1.1
python-dotenv==0.20.0
python-slugify==6.1.2
python-telegram-bot==13.13
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
six==1.16.0
slugify==0.0.1
sniffio==1.2.0
soupsieve==2.3.2.post1
starlette==0.19.1
text-unidecode==1.3
tornado==6.2
typing_extensions==4.3.0
tzdata==2022.1
tzlocal==4.2
uvicorn==0.18.2
uvloop==0.16.0
watchfiles==0.15.0
websockets==10.3
